rename firebase-service.ts to firebase.ts, upgrade deps
