+++ functions/src/functions/listing/update-listing.ts
@@ -1,7 +1,7 @@
 import { ConfirmationResponse } from '../../models/response/confirmation-response';
 import { Collection } from '../../models/collection-name';
 import { Listing, ListingStatus } from '../../models/listing';
-import { Firebase } from '../../services/firebase-service';
+import { Firebase } from '../../services/firebase';
 import { Utils } from '../../utils/utils';
 import { UpdateListingRequest } from '../../models/requests/update-listing-request';
