+++ functions/src/functions/misc/health.ts
@@ -1,3 +1,3 @@
-import { Firebase } from '../../services/firebase-service';
+import { Firebase } from '../../services/firebase';
 
 export const health = Firebase.functions.https.onCall(() => 'ok');
