+++ functions/src/report/createReport.ts
@@ -13,7 +13,7 @@ import {
 	isLoggedIn,
 	isNotBanned,
 } from '../utils';
-import { Firebase } from '../services/firebase-service';
+import { Firebase } from '../services/firebase';
 import { Timestamp } from '../models/timestamp';
 
 const logger = new Logger();
