+++ functions/src/user/updateUserProfile.ts
@@ -2,7 +2,7 @@ import { UserPronoun } from 'types';
 import { ERROR_MESSAGES } from '../constants';
 import Logger from '../Logger';
 import { isLoggedIn, isNotBanned } from '../utils';
-import { Firebase } from '../services/firebase-service';
+import { Firebase } from '../services/firebase';
 
 const logger = new Logger();
