+++ functions/src/utils/fetchListingData.ts
@@ -1,7 +1,8 @@
+import { Firebase } from '../services/firebase';
 import { DEFAULT_GUEST_DISPLAY_NAME, DEFAULT_GUEST_EMAIL } from '../constants';
-import { db } from '../firebase.config';
 import Logger from '../Logger';
 import { ListingData, UserInfo } from '../types';
+
 const logger = new Logger();
 
 export type FetchListingData = (
@@ -16,7 +17,10 @@ const fetchListingData: FetchListingData = async (listingId, invoker) => {
 			invoker.email === DEFAULT_GUEST_EMAIL
 				? 'guest_listings'
 				: 'listings';
-		const docSnap = await db.collection(collection).doc(listingId).get();
+		const docSnap = await Firebase.db
+			.collection(collection)
+			.doc(listingId)
+			.get();
 		return docSnap.data() as ListingData;
 	} catch (error) {
 		logger.error(error);
