+++ functions/src/utils/isLoggedIn.ts
@@ -1,7 +1,7 @@
 import { CallableContext } from 'firebase-functions/v1/https';
 import { ERROR_MESSAGES } from '../constants';
 import Logger from '../Logger';
-import { Firebase } from '../services/firebase-service';
+import { Firebase } from '../services/firebase';
 
 const logger = new Logger();
