+++ functions/src/utils/isNotBanned.ts
@@ -2,7 +2,7 @@ import { UserInfo } from 'types';
 import { ERROR_MESSAGES } from '../constants';
 import Logger from '../Logger';
 import fetchUserInfo from './fetchUserInfo';
-import { Firebase } from '../services/firebase-service';
+import { Firebase } from '../services/firebase';
 
 const logger = new Logger();
