+++ functions/src/wishlist/deleteWishlist.ts
@@ -5,7 +5,7 @@ import {
 } from '../constants';
 import Logger from '../Logger';
 import { isLoggedIn, isNotBanned } from '../utils';
-import { Firebase } from '../services/firebase-service';
+import { Firebase } from '../services/firebase';
 
 const logger = new Logger();
