+++ storage.rules
@@ -13,7 +13,7 @@ service firebase.storage {
       allow read:
         if loggedIn();
       allow write:
-        if loggedIn();
+        if loggedIn()
         && isSeller();
     }
 
@@ -25,7 +25,7 @@ service firebase.storage {
       allow read:
         if loggedIn();
       allow write:
-        if loggedIn();
+        if loggedIn()
         && isSelf();
     }
   }
